Budget refactor
Some changes due to issues. There was an error with updating the correct properties from UserBudget, along with issues with handling the available categories list. I have readjusted the SaveChangeAsync method to be cleaner and more readable at a glance, while being functionally correct. UpdateCategories has been subsidzed as a result of this and has a smaller footprint. With these changes, I added a dictionary to have column names mapped to a category for easy and quick access with SaveChangesAsync.