Budget refactoring (#46)
* Budget refactor

Some changes due to issues. There was an error with updating the correct properties from UserBudget, along with issues with handling the available categories list. I have readjusted the SaveChangeAsync method to be cleaner and more readable at a glance, while being functionally correct. UpdateCategories has been subsidzed as a result of this and has a smaller footprint. With these changes, I added a dictionary to have column names mapped to a category for easy and quick access with SaveChangesAsync.

* Added toast messages; cleanup

Added toast messages to the top of the screen instead of my botched message boxes. Additional budget logic cleanup (no warnings for this page)

* Remove deprecated settings

Originally PoC. These are no longer needed, deleting.

* wip

changes incomplete, moving to other machine. wip

* added modal to add new categories

nuclear option inbound.

Added a modal to the page for when the user wants to add a new category, eliminating any issues with select boxes becoming visually unbound from the class.

* oops missed a spot